move TABLESPACE_URL into tablebase client file
The `TABLESPACE_URL` contain the lichess specific url for endgame
tablebase. We need to use it instead of the lichess one
`https://lichess.org/games`

This constant is located in `berserk/clients/base.py`. This is weird as
this information is specific to tablebase and not useful for other
clients. This logic will be complexe in the future as we will add more
specific urls and overcharge `base.py`.

This commit move this url into `tablebase.py` file, to avoid this