Add opening explorer endpoints
The opening explorer alows lichess users to query statistics about
common moves used in lichess game from a defined position.
It is a great tool to explore opening usage.
https://lichess.org/blog/Vs0xMTAAAD4We4Ey/opening-explorer

It has an API with 3 endpoints:
https://lichess.org/api#tag/Opening-Explorer

It is currently not implemented in berserk

This commit add a new client called ``explore`` which implement on
opening explorer method ``/lichess``