add pgn_in_json parameter for Games.export
Follows a similar pattern to the other pgnInJson parameters.

Note that adding a parameter in the middle of the list is technically a breaking change, but I suspect most people are using kwargs so it won't matter. Let me know if you want me to move the new arg to the end of the arg list.